Decouple PatternMatching dispatcher from 'dispatcher.go' into a separate file, rename dispatcher_test.go to pattern_matching_test.go
